pyproject.toml
[tool.poetry]
name = "qrcode_styled"
version = "0.1.0"
description = "Python port for kozakdenys/qr-code-styling"
authors = ["Bogdan Parfenov <adam.brian.bright@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://cifrazia.com"
repository = "https://github.com/AdamBrianBright/qrcode_styled.git"
keywords = ["cifrazia", "library", "utils", "tools", "qrcode"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: AsyncIO",
    "Natural Language :: English",
    "Natural Language :: Russian",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
packages = [
    { include = "qrcode_styled" }
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/AdamBrianBright/qrcode_styled/issues"
"Contact Author" = "https://vk.com/adam_bright"


[tool.poetry.dependencies]
python = "^3.9"
qrcode = "^6.1"
lxml = { version = "^4.6.3", optional = true }
Pillow = "^8.2.0"

[tool.poetry.extras]
svg = ["lxml"]

[tool.poetry.dev-dependencies]
pytest = "^5.2"
lxml-stubs = "^0.2.0"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
